# [pre-ll] Remove derivative