--- a/src/core/Cargo.toml
+++ b/src/core/Cargo.toml
@@ -30,7 +30,6 @@ path = "src/lib.rs"
 [dependencies]
 bitflags = "1"
 mint = { version = "0.5", optional = true }
-derivative = "1"
 draw_state = "0.8"
 log = "0.4"
 serde = { version = "1", features = ["serde_derive"], optional = true }