--- a/src/core/src/lib.rs
+++ b/src/core/src/lib.rs
@@ -19,8 +19,6 @@
 
 #[macro_use]
 extern crate bitflags;
-#[macro_use]
-extern crate derivative;
 extern crate draw_state;
 extern crate log;