--- a/src/render/Cargo.toml
+++ b/src/render/Cargo.toml
@@ -35,7 +35,6 @@ unstable = []
 
 [dependencies]
 mint = { version = "0.5", optional = true }
-derivative = "1.0"
 draw_state = "0.8"
 gfx_core = { path = "../core", version = "0.9" }
 log = "0.4"