--- a/src/render/src/lib.rs
+++ b/src/render/src/lib.rs
@@ -92,8 +92,6 @@
 extern crate mint;
 
 extern crate log;
-#[macro_use]
-extern crate derivative;
 extern crate draw_state;
 extern crate gfx_core as core;