* Remove crate derivative.
* Fix some warnings: `&Error` -> `&dyn Error`.

After reading about master branch speeding up compile times by removing _derivative_ I wondered if it would help the pre-ll code too. Tbf I can see derivative is quite helpful for the `PhantomData` types where regular `#[derive(...)]` falls down. But it isn't _that_ hard to write out the impls.

It impacts clean builds more than I would expect _(linux, rust 1.38, ryzen 2700x)_.

### Before
```sh
$ hyperfine --max-runs 3 'cargo clean; cargo build --example cube' 
Benchmark #1: cargo clean; cargo build --example cube
  Time (mean ± σ):     34.765 s ±  0.126 s    [User: 248.966 s, System: 12.062 s]
  Range (min … max):   34.666 s … 34.906 s    3 runs
```
### After
```sh
$ hyperfine --max-runs 3 'cargo clean; cargo build --example cube'
Benchmark #1: cargo clean; cargo build --example cube
  Time (mean ± σ):     28.158 s ±  0.343 s    [User: 250.888 s, System: 12.067 s]
  Range (min … max):   27.819 s … 28.504 s    3 runs
```

PR checklist:
- [x] tested examples with the following backends: gl