@alexheretic 2019-10-02
> Thank you! If I understand correctly, this is a non-breaking change.

Yep should function identically.

@kvark Do you think it's worth making a new release with this improvement, or wait for more stuff? I'm happy to raise the v0.18 pr if so.